+++ src/views/mod.rs
@@ -1,2 +1,5 @@
 mod home;
 pub use home::Home;
+
+mod devel;
+pub use devel::Devel;